requirements.txt
pyautogui
pynput
pystray
Pillow
pywebview
